requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
nltk
spacy
faiss-cpu
openpyxl
